+++ src/Rest/AbstractDataProcessor.php
@@ -8,6 +8,7 @@ use Dbp\Relay\CoreBundle\ApiPlatform\State\StateProcessorInterface;
 use Dbp\Relay\CoreBundle\ApiPlatform\State\StateProcessorTrait;
 use Dbp\Relay\CoreBundle\Authorization\AbstractAuthorizationService;
 use Dbp\Relay\CoreBundle\Exception\ApiError;
+use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 
 abstract class AbstractDataProcessor extends AbstractAuthorizationService implements StateProcessorInterface
@@ -31,7 +32,10 @@ abstract class AbstractDataProcessor extends AbstractAuthorizationService implem
     {
         $this->denyOperationAccessUnlessGranted(self::ADD_ITEM_OPERATION);
 
-        $filters = $context[self::FILTERS_CONTEXT_KEY] ?? [];
+        $request = $context['request'] ?? null;
+        assert($request instanceof Request);
+        $filters = $request->query->all();
+
         $this->forbidCurrentUserToAddItemUnlessAuthorized($data, $filters);
 
         return $this->addItem($data, $filters);
@@ -42,7 +46,10 @@ abstract class AbstractDataProcessor extends AbstractAuthorizationService implem
         $this->denyOperationAccessUnlessGranted(self::REPLACE_ITEM_OPERATION);
 
         $currentItem = $context['previous_data'] ?? null;
-        $filters = $context[self::FILTERS_CONTEXT_KEY] ?? [];
+        $request = $context['request'] ?? null;
+        assert($request instanceof Request);
+        $filters = $request->query->all();
+
         $this->forbidCurrentUserToAccessItemUnlessAuthorized(self::REPLACE_ITEM_OPERATION, $currentItem, $filters);
 
         return $this->replaceItem($identifier, $data, $currentItem, $filters);
@@ -53,18 +60,23 @@ abstract class AbstractDataProcessor extends AbstractAuthorizationService implem
         $this->denyOperationAccessUnlessGranted(self::UPDATE_ITEM_OPERATION);
 
         $currentItem = $context['previous_data'] ?? null;
-        $filters = $context[self::FILTERS_CONTEXT_KEY] ?? [];
+        $request = $context['request'] ?? null;
+        assert($request instanceof Request);
+        $filters = $request->query->all();
+
         $this->forbidCurrentUserToAccessItemUnlessAuthorized(self::UPDATE_ITEM_OPERATION, $currentItem, $filters);
 
-        return $this->updateItem($identifier, $data, $currentItem,
-            $context[self::FILTERS_CONTEXT_KEY] ?? []);
+        return $this->updateItem($identifier, $data, $currentItem, $filters);
     }
 
     protected function delete(mixed $identifier, mixed $data, array $context): void
     {
         $this->denyOperationAccessUnlessGranted(self::REMOVE_ITEM_OPERATION);
 
-        $filters = $context[self::FILTERS_CONTEXT_KEY] ?? [];
+        $request = $context['request'] ?? null;
+        assert($request instanceof Request);
+        $filters = $request->query->all();
+
         $this->forbidCurrentUserToAccessItemUnlessAuthorized(self::REMOVE_ITEM_OPERATION, $data, $filters);
 
         $this->removeItem($identifier, $data, $filters);